requirements.txt
gymnasium>=0.29
numpy>=1.24
pyyaml>=6.0
stable-baselines3>=2.0
PyQt6>=6.5
matplotlib>=3.7
fastapi>=0.110
uvicorn[standard]>=0.27
python-jose[cryptography]>=3.3
httpx>=0.27
opencv-python>=4.9
scikit-image>=0.22
torch>=2.2
aiofiles>=23
python-multipart>=0.0.9
slowapi>=0.1.9
ulid-py>=1.1
python-dotenv>=1.0
